requirements.txt
absl-py==1.4.0
astunparse==1.6.3
bayesian-optimization==1.4.3
cachetools==5.3.0
cloudpickle==2.2.1
cloudscraper==1.2.71
flatbuffers==23.5.9
gast==0.4.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
gym==0.26.2
gym-notices==0.0.8
ipython-genutils==0.2.0
jax==0.4.10
kaggle==1.5.13
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
ml-dtypes==0.1.0
munkres==1.1.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
pandas==1.5.3
patsy==0.5.3
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-slugify==8.0.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
TA-Lib==0.4.26
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow-estimator==2.12.0
tensorflow-macos==2.12.0
tensorflow-metal==0.8.0
termcolor==2.3.0
text-unidecode==1.3
webencodings==0.5.1
wrapt==1.14.1
xgboost==1.7.3
